requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.63.0
uvicorn[standard]==0.13.4
gunicorn==20.0.4
python-dotenv==0.17.1
Authlib
httpx
